services/model_img_search/pyproject.toml
[tool.poetry]
name = "model-img-search"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pydantic = "^2.6.4"
aiohttp = "^3.9.3"
numpy = "^1.26.4"
opencv-python-headless = "^4.9.0.80"
python-multipart = "^0.0.9"
torch = "^2.2.0"
torchvision = "^0.17.0"
pillow = "^10.2.0"
openai-clip = "^1.0.1"


[tool.poetry.scripts]
app = "model_img_search.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
